tests/psoc6/dut/i2c: Completed system test machine.i2c.
